Address review comments.
Add docstring comments for some process_* functions in clang/condense.py; remove
a user-side exception in clang/menus.py; restructure clang test_menus.py sample
code to make tests less brittle.
