+++ dxr/plugins/clang/tests/test_menus/code/BaseTypes.h
@@ -1,0 +1,13 @@
+#ifndef BASETYPES_H
+#define BASETYPES_H
+
+class BaseClass {
+public:
+  virtual void virtualFunc() { }
+  virtual void pVirtualFunc() = 0;
+};
+
+struct BaseStruct {
+};
+
+#endif
