+++ dxr/plugins/clang/tests/test_menus/code/DerivedDerivedTypes.h
@@ -1,0 +1,11 @@
+#ifndef DERIVEDDERIVEDTYPES_H
+#define DERIVEDDERIVEDTYPES_H
+
+#include "DerivedTypes.h"
+
+class DerivedDerivedClass : public DerivedClass {
+ public:
+  void virtualFunc() {}
+};
+
+#endif
