+++ dxr/plugins/clang/tests/test_menus/code/DerivedTypes.h
@@ -1,0 +1,15 @@
+#ifndef DERIVEDTYPES_H
+#define DERIVEDTYPES_H
+
+#include "BaseTypes.h"
+
+class DerivedClass : public BaseClass {
+public:
+  void virtualFunc();
+  void pVirtualFunc() { }
+};
+
+struct DerivedStruct : public BaseStruct {
+};
+
+#endif
