+++ dxr/plugins/clang/tests/test_menus/code/extern.c
@@ -27,27 +27,7 @@ namespace Bar = Space;
 
 #define MACRO "polo"
 
-class BaseClass {
-public:
-  virtual void virtualFunc() { }
-  virtual void pVirtualFunc() = 0;
-};
-
-class DerivedClass : public BaseClass {
-public:
-  void virtualFunc();
-  void pVirtualFunc() { }
-};
-
+#include "BaseTypes.h"
+#include "DerivedTypes.h"
 void DerivedClass::virtualFunc() { }
-
-class DerivedDerivedClass : public DerivedClass {
- public:
-  void virtualFunc() {}
-};
-
-struct BaseStruct {
-};
-
-struct DerivedStruct : public BaseStruct {
-};
+#include "DerivedDerivedTypes.h"
